feat add context & likes functionality

+++ src/utils/Api.js
@@ -55,21 +55,38 @@ class Api {
     .then(res => this._getResponse(res));
   }
 
-  addLike(cardId) {
-    return fetch(`${this._baseUrl}${this._id}/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: this._headers
-    })
-    .then(res => this._getResponse(res));
+  toggleLike(cardId, isLiked) {
+    alert(isLiked)
+    if (isLiked === false) {
+      return fetch(`${this._baseUrl}${this._id}/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: this._headers
+      })
+      .then(res => this._getResponse(res));
+    } else {
+      return fetch(`${this._baseUrl}${this._id}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers
+      })
+      .then(res => this._getResponse(res));
+    }
   }
+  //TODO remove methods
+  // addLike(cardId) {
+  //   return fetch(`${this._baseUrl}${this._id}/cards/${cardId}/likes`, {
+  //     method: "PUT",
+  //     headers: this._headers
+  //   })
+  //   .then(res => this._getResponse(res));
+  // }
 
-  removeLike(cardId) {
-    return fetch(`${this._baseUrl}${this._id}/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: this._headers
-    })
-    .then(res => this._getResponse(res));
-  }
+  // removeLike(cardId) {
+  //   return fetch(`${this._baseUrl}${this._id}/cards/${cardId}/likes`, {
+  //     method: "DELETE",
+  //     headers: this._headers
+  //   })
+  //   .then(res => this._getResponse(res));
+  // }
 
   changeAvatar(data) {
     return fetch(`${this._baseUrl}${this._id}/users/me/avatar`, {
